+++ client/requestor/Actions/MoveAction.h
@@ -13,7 +13,7 @@ namespace LM {
         size_t getTargetX() const;
         size_t getTargetY() const;
 
-        void setTarget(uint32_t x, uint32_t y);
+        void setTarget(size_t x, size_t y);
     protected:
         size_t m_TargetX = 0;
         size_t m_TargetY = 0;
